Improvements to new bar
Bar now hides on startup properly, it previously didn't until
we pressed and release super.

The bar used to have issues with layering and appearing under
other windows. To fix this, we now raise the bar every time it
is toggled and when we recieve a new state from bspwm so that
it always appears on top.

Bar now records the state of sxhkd so that we can report
when chords are active. When chords are shown, we also disable
the ability to toggle visibility.

All mapping/unmapping events are now handled within the bar
script so that it can intercept visibility changes for chords.

Fixed an issue with the corner_push script.

Bar now refreshes on resume.

Updated sxhkd to make more liberal use of chords. No more
keys being hogged by launchers.